+++ istrategy.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class IStrategy(ABC):
+    @abstractmethod
+    def search(self, profile):
+        pass
