+++ strategy_fight.py
@@ -1,0 +1,6 @@
+from istrategy import IStrategy
+
+
+class StrategyFight(IStrategy):
+    def search(self, profile):
+        pass
