+++ strategy_love.py
@@ -1,0 +1,6 @@
+from istrategy import IStrategy
+
+
+class StrategyLove(IStrategy):
+    def search(self, profile):
+        pass
